requirements.txt
numpy==1.21.2
PyYAML==6.0
scispacy==0.4.0
spacy==3.1.0
thinc==8.0.10
typer==0.4.0
wasabi==0.8.2
